Move notes module under the notes folder
