+++ src/lib/module/home_notes/notes/home_notes_binding.dart
@@ -1,5 +1,6 @@
 import 'package:get/get.dart';
-import 'package:leafy_launcher/module/home_notes/home_notes_controller.dart';
+
+import 'home_notes_controller.dart';
 
 class HomeNotesBinding extends Bindings {
   @override
