+++ src/lib/module/home_notes/notes/home_notes_page.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/src/widgets/framework.dart';
 import 'package:leafy_launcher/base/page/status_page_base.dart';
-import 'package:leafy_launcher/module/home_notes/home_notes_controller.dart';
 import 'package:leafy_launcher/resources/theme/home_theme.dart';
 import 'package:leafy_launcher/resources/theme/leafy_theme.dart';
+
+import 'home_notes_controller.dart';
 
 class HomeNotesPage extends StatusPageBase<HomeNotesController, HomeTheme> {
   @override
